Port glog to Visual Studio 2015 RTM
glog v0.3.4 has two outstanding issues with the code that prevent it
from compiling on the RTM of Visual Studio 2015 toolchain (and, more
generally, MSVC).

The first is that the function `snprintf` is defined in the version of
stdio.h that ships with the VS2015 toolchain. The fix for this is
relatively straightforward -- we simply define the function only for
versions before the 2015 version.

The second is that the 2015 toolchain will implicitly define a move
constructor for `LogMessage::LogStream`, which in turn will at some
point call a deleted copy constructor for basic_ios. Our solution in
this commit will be to simply explicitly delete it.