Modify the CMakeList.txt
Add a switch ("BUILD_SHARED_LIBS_SWITCH") to control will we build a shared libs. (Default is "ON").

The reason I add this switch is because that sometime we need to build a shared libs, and with this switch we can easily to change to adapt to different situations.
One case:
When compile COLMAP with Ceres (both of them need glog), we must use a shared libs otherwise it will fail. (Although Ceres is ok with a static libs).