remove unnecessary const_cast<char *> in raw_logging.cc
Since afd586a5d5c9, glog has shipped a "const_basename" function
which takes `const char *` and returns a `const char *` pointing to
(a substring of) the input string. This was done as part of a windows
support commit -- previously they used the `basename` function which
takes `char *` and wasn't available on windows.

The `const_cast` may have been necessary at that time, but now it
isn't since we have `const_basename`.

It's also a very dangerous cast because when using the logging
macros, the pointer is usually to the string literal provided by
`__FILE__`, and accessing this pointer as a `char *` yields undefined
behavior, no diagnostic required.

I believe there was no undefined behavior here, because the `char *`
that results is immediately converted to `const char *` as an argument
to `const_basename`. However, this is still a source code improvement.