Fix Bazel build with_gflags=False
as reported by @jondo2010:
https://github.com/google/glog/issues/279#issuecomment-384289681

This was caused by an operator precendence issue: with_gflags=False
would result in copts=[] and the build would not work. This has
presumably never been tested.

Unfortunately, it is not easy to test this in CI with the current
structure. The glog_library() macro can only be used once, as it
creates various rules with fixed names (eg :gen_sh), and must be used at
@com_google_glog//. The only way to use with_gflags=False is to
overwrite the root BUILD file.