Make option that allows passing through vlog on/off
Right now a long sink can't tell if a message is coming from
vlog (or not) so that can make things pretty noisy (especially if
the log sink wishes to ignore such logs).

This adds an option which will pass through to the log sink telling
it if the log originated from a vlog or not (so that it can discard
such messages).