cmake: disable install rules unless standalone build.
If a library is embedded into an outer project, the library install rules
should be disabled not to interfere with the outer project install rules.

A similar setting is used by other projects:

* https://github.com/abseil/abseil-cpp/blob/97ab3dcfd6490434202e4ab00b2eaba9449e42a1/CMakeLists.txt#L60-L64
* https://github.com/google/googletest/blob/7735334a46da480a749945c0f645155d90d73855/CMakeLists.txt#L32

For the same reason, we are unwelcome to modify a default value for a shared
BUILD_SHARED_LIBS option.